tests/wpt/tests/content-security-policy/script-src/tentative/script-url-blocked-report-contains-hash-dedicated-worker.https.html

<!DOCTYPE html>
<html>

<head>
  <title>script-url-blocked-and-sends-hash in dedicated worker</title>
  <script src="/common/get-host-info.sub.js"></script>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/util.js"></script>
</head>

<body>
  <script>
    const scriptUrl = new URL("./support/externalScript.js", document.location).toString();

    promise_test(async t => {
      const expectedHash = await sha256ofURL(scriptUrl);

      return new Promise((resolve, reject) => {
        const w = new Worker("./support/worker-report-hash.js");
        w.onmessage = e => {
          if (e.data.type === 'violation') {
            try {
              assert_equals(e.data.blockedURI, scriptUrl);
              assert_equals(e.data.urlHash.replace(/=+$/, ''), `url-sha256-${expectedHash}`);
              assert_equals(e.data.evalHash, "");
              resolve();
            } catch (err) {
              reject(err);
            }
          }
        };
        w.onerror = (e) => {
          // An error will be triggered because the worker can't import externalScript.js.
          e.preventDefault();
        };
      });
    }, "script-specific violation report contains correct urlHash in dedicated worker");
  </script>
</body>

</html>
